Task-1/microservices/backend/requirements.txt
Flask==2.2.2
psycopg2-binary==2.9.3
Flask-Cors==3.0.0
Werkzeug==2.2.2